+++ examples/graphql_example/entity/src/lib.rs
@@ -1,4 +1,3 @@
 pub mod note;
 
 pub use async_graphql;
-pub use sea_orm;