+++ examples/graphql_example/migration/src/lib.rs
@@ -1,4 +1,4 @@
-pub use sea_schema::migration::*;
+pub use sea_orm_migration::prelude::*;
 
 mod m20220101_000001_create_table;