+++ examples/rocket_example/entity/src/lib.rs
@@ -2,5 +2,3 @@
 extern crate rocket;
 
 pub mod post;
-
-pub use sea_orm;