+++ examples/tonic_example/entity/src/lib.rs
@@ -1,3 +1,1 @@
 pub mod post;
-
-pub use sea_orm;