+++ sea-orm-migration/tests/migrator/m20220118_000001_create_cake_table.rs
@@ -1,0 +1,43 @@
+use sea_orm_migration::prelude::*;
+
+pub struct Migration;
+
+impl MigrationName for Migration {
+    fn name(&self) -> &str {
+        "m20220118_000001_create_cake_table"
+    }
+}
+
+#[async_trait::async_trait]
+impl MigrationTrait for Migration {
+    async fn up(&self, manager: &SchemaManager) -> Result<(), DbErr> {
+        manager
+            .create_table(
+                Table::create()
+                    .table(Cake::Table)
+                    .col(
+                        ColumnDef::new(Cake::Id)
+                            .integer()
+                            .not_null()
+                            .auto_increment()
+                            .primary_key(),
+                    )
+                    .col(ColumnDef::new(Cake::Name).string().not_null())
+                    .to_owned(),
+            )
+            .await
+    }
+
+    async fn down(&self, manager: &SchemaManager) -> Result<(), DbErr> {
+        manager
+            .drop_table(Table::drop().table(Cake::Table).to_owned())
+            .await
+    }
+}
+
+#[derive(Iden)]
+pub enum Cake {
+    Table,
+    Id,
+    Name,
+}