+++ src/entity/active_model.rs
@@ -548,6 +548,12 @@ pub trait ActiveModelTrait: Clone + Debug {
 
         Ok(am)
     }
+
+    /// Return `true` if any field of `ActiveModel` is `Set`
+    fn is_changed(&self) -> bool {
+        <Self::Entity as EntityTrait>::Column::iter()
+            .any(|col| self.get(col).is_set() && !self.get(col).is_unchanged())
+    }
 }
 
 /// A Trait for overriding the ActiveModel behavior
@@ -1087,5 +1093,14 @@ mod tests {
         );
 
         Ok(())
+    }
+
+    #[test]
+    fn test_active_model_is_changed() {
+        let mut fruit: fruit::ActiveModel = Default::default();
+        assert!(!fruit.is_changed());
+
+        fruit.set(fruit::Column::Name, "apple".into());
+        assert!(fruit.is_changed());
     }
 }